Solidify optimistic AV operation
